Convert bypass data to scoreboard tag
